[refactor] resolver-utils 병합, try-catch문 사용
